requirements.txt
streamlit==1.45.1
pillow==11.2.1
pandas==2.2.3
scikit-learn==1.6.1